refactor: 환율 api 관련 파일 위치 이동
