+++ app/src/main/java/com/fintern/ourbudgeting/data/api/ExchangeRateApiService.kt
@@ -1,5 +1,6 @@
 package com.fintern.ourbudgeting.data.api
 
+import com.fintern.ourbudgeting.data.model.ExchangeRate
 import retrofit2.Response
 import retrofit2.http.GET
 import retrofit2.http.Query
